test: fix profile test

+++ src/tests/src/source-helpers.cpp
@@ -5,14 +5,16 @@
 #include "models/profile.h"
 
 
-Profile *makeProfile()
+Profile *makeProfile(bool clean)
 {
-	QFile::remove("tests/resources/blacklist.txt");
-	QFile::remove("tests/resources/favorites.json");
-	QFile::remove("tests/resources/favorites.txt");
-	QFile::remove("tests/resources/ignore.txt");
-	QFile::remove("tests/resources/viewitlater.txt");
-	QFile::remove("tests/resources/wordsc.txt");
+	if (clean) {
+		QFile::remove("tests/resources/blacklist.txt");
+		QFile::remove("tests/resources/favorites.json");
+		QFile::remove("tests/resources/favorites.txt");
+		QFile::remove("tests/resources/ignore.txt");
+		QFile::remove("tests/resources/viewitlater.txt");
+		QFile::remove("tests/resources/wordsc.txt");
+	}
 
 	QFile settings("tests/resources/settings.ini");
 	settings.open(QFile::Truncate | QFile::WriteOnly | QFile::Text);
